requirements.txt
requests==2.32.3
tenacity==9.0.0
ratelimit==2.2.1
tqdm==4.66.4
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
PyYAML==6.0.1
loguru==0.7.2

# dev / optional
black==24.8.0
ruff==0.5.5
pytest==8.2.0
ipykernel==6.29.5


torch>=2.2
torchvision>=0.17
torchaudio>=2.2
pytorch-lightning>=2.3
sentence-transformers>=3.0
transformers>=4.43
scikit-learn>=1.4
torch-geometric>=2.5
pyarrow>=16.0